requirements.txt
Pillow==9.5.0
Requests==2.29.0
streamlit==1.21.0
streamlit_lottie==0.0.3
torch==2.0.0
deta==1.1.0
python-dotenv==1.0.0


gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
psutil  # system resources
PyYAML>=5.3.1
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torchvision>=0.8.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1 # Snyk vulnerability fix